Added Transaction Confirmation page
When clicking on buy/sell, users are routed to confirmation.html. Depending on the purchase type, the page content changes. User must click on the "confirm" on the right, which routes to the buy/sell function and completes it from there. Added function to automatically initialize market hours to 8:00 to 5:00 if the table is empty for any reason. Minor UI changes, like changing padding to dashboards and about page. Deleted unnecessary/redundant functions from app.py.