Timezone implementation
Allows devs to set a timezone that the whole program follows regardless of the machine's local time. Also changes transaction times (on pages, not in the db) to match timezone changes.
If closing time = 5:00 PM, PST = 4:00 PM, and CST = 6:00 PM, setting the timezone to PST would leave the market open to users, and setting it to CST would close it.
In a real-world setting where clients have their timezones dynamically inserted, this would behave strangely. Should consider adjusting this to where market times are set to a single timezone and all timezones sync to the market status.