Updated Sell Page, Abbreviated Volume
The Sell page has been changed from a list of all stocks to a list of the user's portfolio. All instances of the stock field "quantity" has been replaced with "volume". For tables that display stock volume, added a new function which abbreviates large numbers (1,000 to 1k, 2,500,000 to 2.5M, etc.)